made constructor and getters/setters for Star, changed PhysicalBody to return double

+++ src/interfaces/PhysicalBody.java
@@ -1,6 +1,6 @@
 package interfaces;
 
 public interface PhysicalBody extends SpatialEntity {
-    int getMass();
+    double getMass();
     boolean getDestructible();
 }
